feat: #1661 add market expiry to market info
