+++ apps/console-lite/src/app/components/deal-ticket/review-trade.tsx
@@ -9,7 +9,6 @@ import classNames from 'classnames';
 import type { DealTicketMarketFragment } from '@vegaprotocol/deal-ticket';
 import type { OrderSubmissionBody } from '@vegaprotocol/wallet';
 import { SIDE_NAMES } from './side-selector';
-import SimpleMarketExpires from '../simple-market-list/simple-market-expires';
 import { gql, useQuery } from '@apollo/client';
 import type {
   MarketTags,
@@ -17,6 +16,7 @@ import type {
 } from './__generated__/MarketTags';
 import { DealTicketEstimates } from './deal-ticket-estimates';
 import { Side } from '@vegaprotocol/types';
+import { SimpleMarketExpires } from '@vegaprotocol/market-info';
 
 export const MARKET_TAGS_QUERY = gql`
   query MarketTags($marketId: ID!) {
