+++ apps/console-lite/src/app/constants/index.ts
@@ -3,7 +3,6 @@ import { t } from '@vegaprotocol/react-helpers';
 import type { Market } from '@vegaprotocol/market-list';
 
 export const DATE_FORMAT = 'dd MMMM yyyy HH:mm';
-export const EXPIRE_DATE_FORMAT = 'MMM dd';
 
 export const TRADABLE_STATES = {
   [MarketState.STATE_ACTIVE]: true,
