+++ libs/market-info/src/components/index.ts
@@ -1,1 +1,2 @@
 export * from './market-info';
+export * from './market-expires';
