+++ libs/market-info/src/components/market-info/info-market.tsx
@@ -24,6 +24,7 @@ import { useEnvironment } from '@vegaprotocol/environment';
 import { Link as UiToolkitLink } from '@vegaprotocol/ui-toolkit';
 import Link from 'next/link';
 import { AssetDetailsTable, useAssetDataProvider } from '@vegaprotocol/assets';
+import { getMarketExpiryDate } from '../market-expires';
 
 const Links = {
   PROPOSAL_PAGE: ':tokenUrl/governance/:proposalId',
@@ -224,6 +225,9 @@ export const Info = ({ market, onSelect }: InfoProps) => {
       content: (
         <MarketInfoTable
           data={{
+            expiryDate: getMarketExpiryDate(
+              market.tradableInstrument.instrument.metadata.tags
+            ),
             ...market.tradableInstrument.instrument.metadata.tags
               ?.map((tag) => {
                 const [key, value] = tag.split(':');
